Correct Conv2d Remove tests from fixed_prec and add_sharing related to functional APIs
