+++ test/torch/tensors/test_precision.py
@@ -1,7 +1,6 @@
 import pytest
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 
 from syft.frameworks.torch.tensors.interpreters.precision import FixedPrecisionTensor
 
@@ -518,94 +517,6 @@ def test_torch_log_approx(prec_frac, tolerance, workers):
 
     cumsum /= 10
     assert (cumsum.abs() < 1).all()
-
-
-def test_torch_conv2d(workers):
-    bob, alice, james = (workers["bob"], workers["alice"], workers["james"])
-    im = torch.Tensor(
-        [
-            [
-                [[0.5, 1.0, 2.0], [3.5, 4.0, 5.0], [6.0, 7.5, 8.0]],
-                [[10.0, 11.0, 12.0], [13.0, 14.5, 15.0], [16.0, 17.5, 18.0]],
-            ]
-        ]
-    )
-    w = torch.Tensor(
-        [
-            [[[0.0, 3.0], [1.5, 1.0]], [[2.0, 2.0], [2.5, 2.0]]],
-            [[[-0.5, -1.0], [-2.0, -1.5]], [[0.0, 0.0], [0.0, 0.5]]],
-        ]
-    )
-    bias = torch.Tensor([-1.3, 15.0])
-
-    im_fp = im.fix_precision()
-    w_fp = w.fix_precision()
-    bias_fp = bias.fix_precision()
-
-    res0 = torch.conv2d(im_fp, w_fp, bias=bias_fp, stride=1).float_precision()
-    res1 = torch.conv2d(
-        im_fp, w_fp[:, 0:1].contiguous(), bias=bias_fp, stride=2, padding=3, dilation=2, groups=2
-    ).float_precision()
-
-    expected0 = torch.conv2d(im, w, bias=bias, stride=1)
-    expected1 = torch.conv2d(
-        im, w[:, 0:1].contiguous(), bias=bias, stride=2, padding=3, dilation=2, groups=2
-    )
-
-    assert (res0 == expected0).all()
-    assert (res1 == expected1).all()
-
-
-def test_torch_nn_functional_linear():
-    tensor = nn.Parameter(torch.tensor([[1.0, 2], [3, 4]])).fix_prec()
-    weight = nn.Parameter(torch.tensor([[1.0, 2], [3, 4]])).fix_prec()
-
-    result = F.linear(tensor, weight).float_prec()
-
-    expected = torch.tensor([[5.0, 11.0], [11.0, 25.0]])
-
-    assert (result == expected).all()
-
-    tensor = nn.Parameter(torch.tensor([[1.0, -2], [3, 4]])).fix_prec()
-    weight = nn.Parameter(torch.tensor([[1.0, 2], [3, 4]])).fix_prec()
-
-    result = F.linear(tensor, weight).float_prec()
-
-    expected = torch.tensor([[-3.0, -5], [11.0, 25.0]])
-
-    assert (result == expected).all()
-
-    tensor = nn.Parameter(torch.tensor([[1.0, 2], [3, 4]])).fix_prec(precision_fractional=2)
-    weight = nn.Parameter(torch.tensor([[1.0, 2], [3, 4]])).fix_prec(precision_fractional=2)
-
-    result = F.linear(tensor, weight).float_prec()
-
-    expected = torch.tensor([[5.0, 11.0], [11.0, 25.0]])
-
-    assert (result == expected).all()
-
-
-def test_torch_nn_functional_dropout(workers):
-    # Only for precision tensor
-    a = torch.rand((20, 20))
-    x = a.fix_prec()
-
-    train_output = F.dropout(x, p=0.5, training=True, inplace=False)
-    assert (train_output.float_prec() == 0).sum() > 0
-
-    test_output = F.dropout(x, p=0.5, training=False, inplace=False)
-    # should return the same input
-    assert ((test_output == x).float_prec() == 1).all()
-
-    # For AST with precision
-    bob, alice, james = (workers["bob"], workers["alice"], workers["james"])
-    x = a.fix_prec().share(alice, bob, crypto_provider=james)
-
-    train_output = F.dropout(x, p=0.5, training=True, inplace=False)
-    assert (train_output.get().float_prec() == 0).sum() > 0
-
-    test_output = F.dropout(x, p=0.5, training=False, inplace=False)
-    assert ((test_output == x).get().float_prec() == 1).all()
 
 
 def test_operate_with_integer_constants():
